services/processing/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sidekick-processing"
version = "0.1.0"
description = "Acquisition and processing workers (Phase 3)"
requires-python = ">=3.12"
dependencies = [
    "sidekick-core",
    "typer>=0.24.1",
    "pypdf>=5.0",
    "python-dotenv>=1.0",
    "torch>=2.10",
    "whisperx>=3.1.1",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
]

[project.scripts]
sidekick-process = "sidekick.processing.cli:main"

[tool.uv.sources]
sidekick-core = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/sidekick"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
